# src/swagger.yaml
swagger: '2.0'
info:
  title: DEEL BACKEND TASK
  description: Api endpoints
  version: 0.1.0
schemes:
  - http
basePath: /
produces:
  - application/json
paths:
...
  /jobs/unpaid:
    get:
      summary: Return all unpaid jos for profile
      tags:
        - Jobs
      parameters:
        - name: profile_id
          type: number
          in: header
          required: true
      responses:
        '200':
          description: Jobs
          schema:
            type: array
            items:
              $ref: '#/definitions/Job'
        '403':
          description: Forbidden
...
  /balances/deposit/{userId}:
    post:
      summary: Deposit funds to client`s account
      tags:
        - Balances
      parameters:
        - name: userId
          type: number
          in: path
          required: true
        - name: amount
          schema:
            type: object
            properties:
              amount:
                type: number
                description: Amount to deposit
            required:
              - amount
          in: body
          required: true
      responses:
        '201':
          description: Created
        '400':
          description: Bad request
...
definitions:
...
  Job:
    type: object
    properties:
      id:
        type: number
      description:
        type: string
      price:
        type: number
      paid:
        type: boolean
      paymentDate:
        type: string
      createdAt:
        type: string
      updatedAt:
        type: string
      ContractId:
        type: number